% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-newdata.R
\name{make_newdata}
\alias{make_newdata}
\alias{make_newdata.default}
\title{Construct a data frame suitable for prediction}
\usage{
make_newdata(x, ...)

\method{make_newdata}{default}(x, ...)
}
\arguments{
\item{x}{A data frame (or object that inherits from \code{data.frame}).}

\item{...}{Covariate specifications (expressions) that will be evaluated
by looking for variables in \code{x} (or \code{data}). Must be of the form \code{z = f(z)}
where \code{z} is a variable in the data set \code{x} and \code{f} a known
function that can be usefully applied to \code{z}. See examples below.}
}
\description{
Given a data set, returns a data set that can be used
as \code{newdata} argument in a call to \code{predict} and similar functions.
}
\details{
Depending on the class of \code{x}, mean or modus values will be
used for variables not specified in ellipsis. If x is an object that inherits
from class \code{ped}, useful data set completion will be attempted depending
on variables specified in ellipsis.
}
\examples{
library(dplyr)
iris \%>\% make_newdata()
iris \%>\% make_newdata(Sepal.Length=c(5))
iris \%>\% make_newdata(Sepal.Length=seq_range(Sepal.Length, 3),
 Sepal.Width=c(2, 3))
iris \%>\% make_newdata(Sepal.Length=seq_range(Sepal.Length, 3),
 Species=unique(Species), Sepal.Width=c(2, 3))
# mean/modus values of unspecified variables are calculated over whole data
iris \%>\% make_newdata(Species=unique(Species))
iris \%>\% group_by(Species) \%>\% make_newdata()
# You can also pass a part of the data sets as data frame to make_newdata
purrr::cross_df(list(Sepal.Length = c(0, 500, 1000),
 Species = c("setosa", "versicolor"))) \%>\%
  make_newdata(x=iris)
}
