% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPlotData.R
\name{tidy_s}
\alias{tidy_re}
\alias{tidy_s}
\title{Extract 1d smooth objects in tidy data format.}
\usage{
tidy_s(po, keep = c("x", "fit", "se", "xlab", "ylab"), se = TRUE)

tidy_re(po, keep = c("fit", "main", "xlab", "ylab"))
}
\arguments{
\item{po}{A list of plot obects as returned from \code{\link[mgcvtools]{get_plotinfo}}.}

\item{keep}{A vector of variables to keep.}
}
\description{
Extract 1d smooth objects in tidy data format.

Extract random effects objects in tidy data format.
}

